Rename CLI project path to src/AiCLI and update references
