+++ AiLang.slnx
@@ -1,6 +1,6 @@
 <Solution>
   <Folder Name="/src/">
-    <Project Path="src/AiLang.Cli/AiLang.Cli.csproj" />
+    <Project Path="src/AiCLI/AiCLI.csproj" />
     <Project Path="src/AiLang.Core/AiLang.Core.csproj" />
     <Project Path="src/AiVM.Core/AiVM.Core.csproj" />
   </Folder>
