+++ src/AiCLI/AiCLI.csproj
