+++ src/AiCLI/CliHttpServe.cs
